backend/main_server/python/requirements.txt
openai-whisper
torch
numpy
scipy